agents/requirements.txt
google-cloud-aiplatform[adk,agent_engines]==1.110.0